requirements.txt
flask_cors==5.0.1
flask-socketio>=5.3.0
Flask
anthropic>=0.42.0
claude-agent-sdk
python-dotenv
pyyaml